examples/wkmkclext: wkmkclext-simple
I wasn't able to debug the failures with the mkcl-dumped
libwkmkclext-library.so.  So the older approach is implemented as an
alternative.

A webkit webextension module called wkmkclext-simple.so is produced
from C.  wkmkclext.lisp still defines the webkit initialization
function as a cffi callback but this is not exported through ELF.
Instead we arrange for this function to be called by the C-defined
`webkit_web_extension_initialize'.  This is done by passing the
callback address to C when wkmkclext.lisp is loaded when mkcl is
initialized through the mkclplug module loading facility.

* examples/wkmkclext/Makefile: (wkmkclext-simple.so): new target.  a
webkit-web-extension loadable module built with C.

* examples/wkmkclext/wkext-simple.c:
(register_init1): takes a single argument to store the address of a
lisp callback function which initializes the webkit side and stashes
it. (webkit_web_extension_initialize): call the stashed function.

wkmkclext-simple.o doesn't need to know about any lisp since there
isn't any initialization. Generate it from wkext-simple.o by defining
WKEXT_APPNAME='"wkmkclext"'

There is need for a check-init-simple.c since there is no need for CL
initialization. Only the webkit extension has to be initialized.
GModuleCheck facility does just that. handle stub init via -DWK_MKCL

* examples/wkmkclext/wkmkclext.lisp: conditionalize on :wkmkclext-simple
in *features*.

;madhu 221211 fixed by arranging to set the environment variable
WKEXTSIMPLE (in wkext-simple.c:(init_wkmkclext)), and using this to
set the lisp feature when loading initrc

;madhu 221229

* examples/wkmkclext/wkmkclext.system: we try to detect problems with
compile-time mismatch of the :wkmkclext-simple feature, but don't take
any chances: if :wkmkclext-simple is in features don't compile
wkmkclext.lisp always load it.